+++ frontend/src/app/dashboard/page.tsx
@@ -7,7 +7,7 @@ import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Progress } from "@/components/ui/progress";
 import { MapPin, Award, Target, ChevronRight } from 'lucide-react';
-import { Footer } from './Footer';
+import { Footer } from '../components/Footer';
 
 export default function DashboardPage() {
   // Mock data - replace with actual data fetching logic
